pyproject.toml
[tool.poetry]
name = "teste-drf"
version = "0.1.0"
description = ""
authors = ["Henrique Faria <hgf777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
markdown = "^3.7"
pygments = "^2.19.1"
djangorestframework-simplejwt = "^5.4.0"
django-rest-knox = "^5.0.2"
django-oauth-toolkit = "^3.0.1"
gunicorn = "^23.0.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
